Work on ex 2.20 half way
Work on adding notification prop and style class.
Get stuck on debugging something silly. React fundamentally runs alot of
renders, but I thought that was a problem. Spent a lot of time on
something that isn't an issue (but I learned stuff, so that's good!)

It should be really simple to wrap 2.20 up, I did the hard part already.