Finish reading 3b
Finish reading part 3b, and have the notes application deployed to
Heroku. Working on the part 3b exercises.

Have to move projects out of the monorepo for now, although I have plans
to put them back in and let Heroku handle the mono repo subprojects
based on that last link in the notes.